Notifier/requirements.txt
beautifulsoup4==4.11.1
flask
flask[async]
python-telegram-bot==20.0
requests==2.28.1
requests-oauthlib==1.3.1
oauthlib==3.2.0
deta